Added last part of dotnet conf - Missing sessions will not be included anymore
As also mentioned in the Readme.md of this year, I will not be including missing sessions from the point mentioned in the readme.
This is because I think they offer little additional value over the already watched sessions as well as taking a considerable amount of time to watch those.